+++ client/src/configurator/BlockTypeSettings.js
@@ -9,6 +9,7 @@ const _defaults = {
   id: null,
   sortOrder: 0,
   fieldLayoutId: null,
+  fieldLayoutConfig: null,
   name: '',
   handle: '',
   description: '',
@@ -45,6 +46,7 @@ export default Settings.extend({
     this._childBlocks = settings.childBlocks
     this._id = settings.id
     this._fieldLayoutId = settings.fieldLayoutId
+    this._fieldLayoutConfig = settings.fieldLayoutConfig
     this._errors = settings.errors
     this._js = settings.js
     this._settingsChildBlockTypes = settings.childBlockTypes
@@ -149,6 +151,10 @@ export default Settings.extend({
 
   getFieldLayoutId () {
     return this._fieldLayoutId
+  },
+
+  getFieldLayoutConfig () {
+    return Object.assign({}, this._fieldLayoutConfig)
   },
 
   isNew () {